Store OpenGL texture staging areas as `uint32_t` vectors

+++ src/gui/render/render_backend.h
@@ -78,10 +78,10 @@ public:
 	// If a renderer implements a double buffering scheme, this call should
 	// return a pointer to the current render buffer.
 	//
-	// `pitch_out` is the number of bytes used to store a single row of
-	// pixel data (can be larger than actual width).
+	// `pitch_out` is the number of bytes used to store a single row of pixel
+	// data, including optional padding bytes at the end of the row.
 	//
-	virtual void StartFrame(uint8_t*& pixels_out, int& pitch_out) = 0;
+	virtual void StartFrame(uint32_t*& pixels_out, int& pitch_out) = 0;
 
 	// Called at the end of every frame. There is a matching EndUpdate()
 	// call for every StartUpdate() call.
